requirements.txt
psycopg2-binary==2.9.1
SQLAlchemy==1.4.31
schedule==1.1.0
python-dotenv==0.21.0
python-telegram-bot==13.14
